Auth & Sign-In Actions Upgrade

+++ src/store/actions/user.js
@@ -1,5 +1,6 @@
 import { createAsyncThunk } from "@reduxjs/toolkit"
 import axios from 'axios'
+import Swal from "sweetalert2"
 
 const register = createAsyncThunk('register', async payload => {
   try {
@@ -25,44 +26,59 @@ const register = createAsyncThunk('register', async payload => {
   }
 })
 
-const sign_in = createAsyncThunk('sign_in', async payload => {
+const sign_in = createAsyncThunk('sign_in', async (payload, {rejectWithValue}) => {
   try {
     let {email,password} = payload
-    const user = await axios
-      .post('http://localhost:3000/api/user/login', {
-        email: email,
-        password: password
+    const response = await axios.post('http://localhost:3000/api/user/login', {
+      email: email,
+      password: password,
+    })
+    if (response.data.user) {
+      // console.log('User successfully logged in')
+      const Toast = Swal.mixin({
+        toast: true,
+        position: 'top-end',
+        showConfirmButton: false,
+        showCloseButton: true,
+        timer: 4000,
+        didOpen: (toast) => {
+          toast.addEventListener('mouseenter', Swal.stopTimer)
+          toast.addEventListener('mouseleave', Swal.resumeTimer)
+        }
+      })        
+      Toast.fire({
+        icon: 'success',
+        title: 'Successful login'
       })
-      .then(response => {
-        console.log('User successfully logged in')
-        localStorage.setItem('token', response.data.token)
-        return response.data.user
-      })
-      .catch(error => console.log(error))
-    return {user: user}
-  }
-  catch (error) {
-    console.log(error.message)
+      localStorage.setItem('token', response.data.token)
+      return response.data.user
+    } else {
+      console.log('Login failed');
+      return rejectWithValue({message: 'Login failed'})
+    }
+  } catch (error) {
+    console.log('Login error:', error)
+    return rejectWithValue(error)
   }
 })
 
-const authenticate = createAsyncThunk('authenticate', async () => {
+const authenticate = createAsyncThunk('authenticate', async (_, {rejectWithValue}) => {
   try {
     let token = localStorage.getItem('token')
-    let user = await axios
-      .post('http://localhost:3000/api/user/auth', null, {
-        headers: {'Authorization': 'Bearer ' + token}
-      })
-      .then(response => {
-        console.log('User successfully authenticated')
-        localStorage.setItem('token', response.data.token)
-        return response.data.user
-      })
-      .catch(error => console.log(error))
-    return {user: user}
-  } 
-  catch (error) {
-    console.log(error.message)
+    let response = await axios.post('http://localhost:3000/api/user/auth', null, {
+      headers: { 'Authorization': 'Bearer ' + token }
+    })
+    if (response.data.user) {
+      console.log('User successfully authenticated')
+      localStorage.setItem('token', response.data.token)
+      return response.data.user
+    } else {
+      console.log('Authentication failed')
+      return rejectWithValue({message: 'Authentication failed'})
+    }
+  } catch (error) {
+    console.log('Authentication error: ', error)
+    return rejectWithValue(error);
   }
 })
 
@@ -71,6 +87,7 @@ const sign_out = createAsyncThunk('sign_out', async () => {
     await axios
     .post('http://localhost:3000/api/user/logout')
     .then(localStorage.removeItem('token'))
+    .then(console.log('Successful logout'))    
   }
   catch (error) {
     console.log(error.message)
